remove old price range filter
